backend/requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
antlr4-python3-runtime
anyio
asgiref
attrs
backoff
bcrypt
beautifulsoup4
build
cachetools
certifi
cffi
chardet
charset-normalizer
chroma-hnswlib
chromadb
click
coloredlogs
contourpy
cryptography
cycler
dataclasses-json
Deprecated
distro
dnspython
durationpy
effdet
email_validator
emoji
eval_type_backport
fastapi
fastapi-cli
filelock
filetype
flatbuffers
fonttools
frozenlist
fsspec
google-api-core
google-auth
google-cloud-vision
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
html5lib
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
iopath
Jinja2
jiter
joblib
jsonpatch
jsonpath-python
jsonpointer
kiwisolver
kubernetes
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langdetect
langgraph
langgraph-checkpoint
langgraph-sdk
langsmith
layoutparser
lxml
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
mmh3
monotonic
mpmath
msgpack
multidict
mypy-extensions
ndjson
nest-asyncio
networkx
nltk
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
olefile
omegaconf
onnx
onnxruntime
openai
opencv-python
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pandas
pdf2image
pdfminer.six
pdfplumber
pi_heif
pikepdf
pillow
portalocker
posthog
propcache
proto-plus
protobuf
psutil
pyasn1
pyasn1_modules
pycocotools
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
pypandoc
pyparsing
pypdf
pypdfium2
PyPika
pyproject_hooks
python-dateutil
python-docx
python-dotenv
python-iso639
python-magic
python-multipart
python-oxmsg
python-pptx
pytz
PyYAML
RapidFuzz
regex
requests
requests-oauthlib
requests-toolbelt
rich
rich-toolkit
rsa
safetensors
scipy
setuptools
shellingham
six
sniffio
soupsieve
SQLAlchemy
starlette
sympy
tenacity
tiktoken
timm
tokenizers
torch
torchvision
tqdm
transformers
triton
typer
typing-inspect
typing_extensions
tzdata
unstructured
unstructured-client
unstructured-inference
unstructured.pytesseract
urllib3
uvicorn
uvloop
watchfiles
webencodings
websocket-client
websockets
wheel
wrapt
XlsxWriter
yarl
zipp
